refactor: rename imagePreviewStore to nodeOutputStore
Align filename with the store it houses (useNodeOutputStore / Pinia ID
'nodeOutput'). Pure mechanical rename — no logic changes.

Amp-Thread-ID: https://ampcode.com/threads/T-019cbc5f-bf57-7219-b27a-55708f7c193a
