+++ src/composables/maskeditor/useMaskEditorSaver.ts
@@ -1,6 +1,6 @@
 import { useMaskEditorDataStore } from '@/stores/maskEditorDataStore'
 import { useMaskEditorStore } from '@/stores/maskEditorStore'
-import { useNodeOutputStore } from '@/stores/imagePreviewStore'
+import { useNodeOutputStore } from '@/stores/nodeOutputStore'
 import type {
   EditorOutputData,
   EditorOutputLayer,
